charms/kfp-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.7.1
    # via httpcore
attrs==23.1.0
    # via jsonschema
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charmed-kubeflow-chisme==0.2.0
    # via -r requirements.in
charset-normalizer==3.2.0
    # via requests
cosl==0.0.11
    # via -r requirements.in
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
exceptiongroup==1.1.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via lightkube
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.0.1
    # via jsonschema
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jsonschema==4.17.3
    # via serialized-data-interface
lightkube==0.14.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
lightkube-models==1.27.1.4
    # via lightkube
markupsafe==2.1.3
    # via jinja2
oci-image==1.0.0
    # via -r requirements.in
ops==2.5.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
    #   cosl
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pkgutil-resolve-name==1.3.10
    # via jsonschema
pyrsistent==0.19.3
    # via jsonschema
pyyaml==6.0.1
    # via
    #   cosl
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via serialized-data-interface
ruamel-yaml==0.17.32
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.2.2
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
typing-extensions==4.11.0
    # via cosl
urllib3==2.0.4
    # via requests
websocket-client==1.6.1
    # via ops
zipp==3.16.2
    # via importlib-resources
